# DSell_API_OpenAPI.json
{
    "openapi":  "3.0.3",
    "info":  {
                 "title":  "DSell 直销管理系统 API",
                 "version":  "1.0.0",
                 "description":  "Base URL: http://localhost:8080\n认证方式: Header 中携带 token (登录接口返回)"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080"
                    }
                ],
    "security":  [
                     {
                         "token":  [

                                   ]
                     }
                 ],
    "components":  {
                       "securitySchemes":  {
                                               "token":  {
                                                             "type":  "apiKey",
                                                             "in":  "header",
                                                             "name":  "token"
                                                         }
                                           },
                       "schemas":  {
                                       "LoginDTO":  {
                                                        "properties":  {
                                                                           "password":  {
                                                                                            "type":  "string",
                                                                                            "example":  "admin123"
                                                                                        },
                                                                           "username":  {
                                                                                            "description":  "管理员用户名",
                                                                                            "example":  "admin",
                                                                                            "type":  "string"
                                                                                        }
                                                                       },
                                                        "type":  "object",
                                                        "required":  [
                                                                         "username",
                                                                         "password"
                                                                     ]
                                                    },
...
                                       "CaptchaSendRequest":  {
                                                                  "type":  "object",
                                                                  "required":  [
                                                                                   "type",
                                                                                   "scene",
                                                                                   "target"
                                                                               ],
                                                                  "properties":  {
                                                                                     "type":  {
                                                                                                  "type":  "string",
                                                                                                  "enum":  [
                                                                                                               "sms",
                                                                                                               "email"
                                                                                                           ]
                                                                                              },
                                                                                     "scene":  {
                                                                                                   "type":  "string",
                                                                                                   "enum":  [
                                                                                                                "login",
                                                                                                                "register",
                                                                                                                "bind",
                                                                                                                "forget"
                                                                                                            ]
                                                                                               },
                                                                                     "target":  {
                                                                                                    "type":  "string",
                                                                                                    "example":  "13800000001"
                                                                                                }
                                                                                 }
                                                              },
                                       "CaptchaVerifyRequest":  {
                                                                    "type":  "object",
                                                                    "required":  [
                                                                                     "type",
                                                                                     "scene",
                                                                                     "target",
                                                                                     "code"
                                                                                 ],
                                                                    "properties":  {
                                                                                       "type":  {
                                                                                                    "type":  "string",
                                                                                                    "enum":  [
                                                                                                                 "sms",
                                                                                                                 "email"
                                                                                                             ]
                                                                                                },
                                                                                       "scene":  {
                                                                                                     "type":  "string",
                                                                                                     "enum":  [
                                                                                                                  "login",
                                                                                                                  "register",
                                                                                                                  "bind",
                                                                                                                  "forget"
                                                                                                              ]
                                                                                                 },
                                                                                       "target":  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                       "code":  {
                                                                                                    "type":  "string",
                                                                                                    "example":  "888888"
                                                                                                }
                                                                                   }
                                                                },
...
                                   }
                   },
...
    "paths":  {
                  "/api/auth/login":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "Auth"
                                                                ],
                                                       "summary":  "管理端登录 (用户名+密码)",
                                                       "security":  [

                                                                    ],
                                                       "requestBody":  {
                                                                           "required":  true,
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/LoginDTO"
                                                                                                                               }
                                                                                                                }
                                                                                       }
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "返回token和用户信息"
                                                                                 }
                                                                     }
                                                   }
                                      },
...
                  "/api/captcha/send":  {
                                            "post":  {
                                                         "tags":  [
                                                                      "Captcha"
                                                                  ],
                                                         "summary":  "发送验证码",
                                                         "security":  [

                                                                      ],
                                                         "requestBody":  {
                                                                             "required":  true,
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/CaptchaSendRequest"
                                                                                                                                 }
                                                                                                                  }
                                                                                         }
                                                                         }
                                                     }
                                        },
                  "/api/captcha/verify":  {
                                              "post":  {
                                                           "tags":  [
                                                                        "Captcha"
                                                                    ],
                                                           "summary":  "校验验证码",
                                                           "security":  [

                                                                        ],
                                                           "requestBody":  {
                                                                               "required":  true,
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/CaptchaVerifyRequest"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                       }
                                          },
...
                  "/api/auth/social/url":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "Social"
                                                                    ],
                                                           "summary":  "获取社交授权URL",
                                                           "security":  [

                                                                        ],
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "type",
                                                                                  "in":  "query",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              },
                                                                              {
                                                                                  "name":  "redirectUri",
                                                                                  "in":  "query",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              }
                                                                          ]
                                                       }
                                           },
                  "/api/auth/social/login":  {
                                                 "post":  {
                                                              "tags":  [
                                                                           "Social"
                                                                       ],
                                                              "summary":  "社交登录",
                                                              "security":  [

                                                                           ],
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "type":  "object",
                                                                                                                                          "properties":  {
                                                                                                                                                             "type":  {
                                                                                                                                                                          "type":  "string"
                                                                                                                                                                      },
                                                                                                                                                             "code":  {
                                                                                                                                                                          "type":  "string"
                                                                                                                                                                      },
                                                                                                                                                             "state":  {
                                                                                                                                                                           "type":  "string"
                                                                                                                                                                       }
                                                                                                                                                         }
                                                                                                                                      }
                                                                                                                       }
                                                                                              }
                                                                              }
                                                          }
                                             },
                  "/api/auth/social/bind":  {
                                                "post":  {
                                                             "tags":  [
                                                                          "Social"
                                                                      ],
                                                             "summary":  "绑定社交账号",
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "type":  "object",
                                                                                                                                         "properties":  {
                                                                                                                                                            "type":  {
                                                                                                                                                                         "type":  "string"
                                                                                                                                                                     },
                                                                                                                                                            "code":  {
                                                                                                                                                                         "type":  "string"
                                                                                                                                                                     }
                                                                                                                                                        }
                                                                                                                                     }
                                                                                                                      }
                                                                                             }
                                                                             }
                                                         }
                                            },
                  "/api/auth/social/unbind":  {
                                                  "post":  {
                                                               "tags":  [
                                                                            "Social"
                                                                        ],
                                                               "summary":  "解绑社交账号",
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "type":  "object",
                                                                                                                                           "properties":  {
                                                                                                                                                              "type":  {
                                                                                                                                                                           "type":  "string"
                                                                                                                                                                       }
                                                                                                                                                          }
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               }
                                                           }
                                              }
              }
}